fix: Fix invalid padding for some cases of ansible vault context action
In some cases events are sent in the same line for the process handler when the outputs get longer. Such cases lead to incorrectly padded vault files, which in some versions of ansible and with some tools can lead to errors.
Besides that it changes the ansible-vault output which is not intended, to make sure it is as portable and "near to the cli" as possible.